+++ platform/fabric/core/generic/membership.go
@@ -7,10 +7,30 @@ SPDX-License-Identifier: Apache-2.0
 package generic
 
 import (
+	api2 "github.com/hyperledger-labs/fabric-smart-client/platform/view/driver"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
 	"github.com/hyperledger/fabric/msp"
+	"github.com/pkg/errors"
 
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/driver"
 )
+
+func (c *Channel) IsValid(identity view.Identity) error {
+	id, err := c.MSPManager().DeserializeIdentity(identity)
+	if err != nil {
+		return errors.Wrapf(err, "failed deserializing identity [%s]", identity.String())
+	}
+
+	return id.Validate()
+}
+
+func (c *Channel) GetVerifier(identity view.Identity) (api2.Verifier, error) {
+	id, err := c.MSPManager().DeserializeIdentity(identity)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed deserializing identity [%s]", identity.String())
+	}
+	return id, nil
+}
 
 // GetMSPIDs retrieves the MSP IDs of the organizations in the current Channel
 // configuration.
