--- platform/fabric/core/generic/signer.go
@@ -1,13 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package generic
-
-type SerializableSigner interface {
-	Sign(message []byte) ([]byte, error)
-
-	Serialize() ([]byte, error)
-}
